demo-backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.11.2
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
